+++ mod-source-record-storage-server/src/main/java/org/folio/services/handlers/actions/MarcAuthorityUpdateModifyEventHandler.java
@@ -29,6 +29,11 @@ public class MarcAuthorityUpdateModifyEventHandler extends AbstractUpdateModifyE
   }
 
   @Override
+  protected boolean isHridFillingNeeded() {
+    return false;
+  }
+
+  @Override
   public String getPostProcessingInitializationEventType() {
     return DI_SRS_MARC_AUTHORITY_RECORD_MODIFIED_READY_FOR_POST_PROCESSING.value();
   }