+++ NEWS.md
@@ -4,7 +4,10 @@
 * [MODSOURMAN-724](https://issues.folio.org/browse/MODSOURMAN-724) SRM does not process and save error records
 * [MODSOURCE-477](https://issues.folio.org/browse/MODSOURCE-477) Configure job to delete authority records 
 
-## 2022-03-xx v5.3.1-SNAPSHOT
+## 2022-03-xx v5.3.2-SNAPSHOT
+* [MODSOURCE-489](https://issues.folio.org/browse/MODSOURCE-489) Records that are overlaid multiple times via Inventory Single Record Import can't be overlaid after the second time
+
+## 2022-03-25 v5.3.1
 * [MODSOURCE-482](https://issues.folio.org/browse/MODSOURCE-482) One-record OCLC (Create) Data Import takes over 9 seconds
 
 ## 2021-02-22 v5.3.0